Force all tests to pass (#128)
Apparently at some point we decided to be ok with cases in `corpus.py` to NOT pass in `make test`, i.e. not generate a solution given the parameters of the setup. Also, we set the parameter `max_stack_depth=20` when running tests to speed them up. All this led to a whole bunch of tests silently failing without ever producing the desired resolution.

Here this is changed again.

* All test cases in `corpus.py` must yield the desired resolution when running `make test` - else tests fail
* To achieve this I made the following changes to the test runner
  - max_stack_depth = 100
  - Use the scoring model - not the DummyScorer; but only when running tests, NOT when generating training samples
  - Allow for the latter distinction
  - Still, some tests were failing. I had to add extra "weight" for YYYYMMDD - YYYYMMDD intervals and slightly changed
    a very specific case to avoid needing an even higher `max_stack_depth`

* nits